requirements.txt
pyserial>3
readline
tk
tqdm
